requirements.txt
africastalking==1.2.9
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aioquic==1.2.0
aiosignal==1.3.2
anyio==3.5.0
APScheduler==3.6.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.7.1
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
cachetools==4.2.2
certifi==2021.10.8
cffi==1.17.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.8
colorama==0.4.6
config==0.4.2
crypto==1.4.1
cryptography==44.0.2
dnspython==2.7.0
ExtProxy==1.0.3
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
html5lib==1.0.1
httpcore==0.17.3
httpx==0.24.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.3
itsdangerous==2.2.0
Jinja2==3.1.6
kaitaistruct==0.10
ldap3==2.9.1
loguru==0.6.0
lolcat==1.4
MarkupSafe==3.0.2
mechanize==0.4.10
mitmproxy==11.1.3
mitmproxy_linux==0.11.5
mitmproxy_rs==0.11.5
MouseInfo==0.1.3
msgpack==1.1.0
multidict==6.2.0
Naked==0.1.32
netifaces==0.11.0
paramiko==3.5.1
passlib==1.7.4
phonenumbers==8.10.2
phpserialize==1.3
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.22.0
pydantic==1.9.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.10.1
pylsqpack==0.3.20
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pysnmp==7.1.17
PySocks==1.7.1
python-telegram-bot==20.3
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
PyYAML==6.0.2
questionary==2.1.0
requests==2.32.2
rfc3986==1.5.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scapy==2.6.1
schema==0.7.7
selenium==3.141.0
service-identity==24.2.0
setuptools==78.1.0
shellescape==3.8.1
six==1.17.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.6
termcolor==2.5.0
tornado==6.1
truecallerpy==1.0.3
twilio==9.5.1
typing_extensions==4.13.1
tzlocal==5.3.1
urllib3==2.3.0
urwid==2.6.16
user_agent==0.1.10
uuid==1.30
validators==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wget==3.2
wheel==0.45.1
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
